testing/web-platform/tests/resource-timing/buffer-full-inspect-buffer-during-callback.html

<!DOCTYPE HTML>
<html>
<head onload>
<meta charset="utf-8" />
<title>This test validates the buffer doesn't contain more entries than it should inside onresourcetimingbufferfull callback.</title>
<link rel="author" title="Intel" href="http://www.intel.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/#performanceresourcetiming"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/append.js"></script>
</head>
<body>
<script>
    const t = async_test("Verify that inspecting the resource timing buffer during resourcetimingbufferfull call doesn't exceed the limit.");
    let resource_timing_buffer_size = 2;
    performance.clearResourceTimings();
    var resize = function() {
        assert_equals(performance.getEntriesByType("resource").length, resource_timing_buffer_size, "resource timing buffer in resourcetimingbufferfull is the size of the limit");
        ++resource_timing_buffer_size;
        performance.setResourceTimingBufferSize(resource_timing_buffer_size);
        xhrScript("resources/empty.js?xhr");
        assert_equals(performance.getEntriesByType("resource").length, resource_timing_buffer_size - 1, "A sync request was not added to the primary buffer just yet, because it is full");
        ++resource_timing_buffer_size;
        performance.setResourceTimingBufferSize(resource_timing_buffer_size);
    }
    performance.addEventListener('resourcetimingbufferfull', t.step_func(resize));
    performance.setResourceTimingBufferSize(resource_timing_buffer_size);
    // Scripts appended in JS to ensure setResourceTimingBufferSize is called before.
    // These resources gets buffered in the resource timing entry buffer.
    appendScript('resources/empty.js');
    appendScript('resources/empty.js?second');
    // This resource overflows the entry buffer, and goes into the secondary buffer.
    appendScript('resources/empty_script.js');
    window.onload = t.step_func_done(()=>{
        let entries = performance.getEntriesByType('resource');
        assert_equals(entries.length, resource_timing_buffer_size,
                      'All 4 entries should be stored in resource timing buffer.');
        assert_true(entries[0].name.includes('empty.js'), "empty.js is in the entries buffer");
        assert_true(entries[1].name.includes('empty.js?second'), "empty.js?second is in the entries buffer");
        assert_true(entries[2].name.includes('empty_script.js'), "empty_script.js is in the entries buffer");
        assert_true(entries[3].name.includes('empty.js?xhr'), "empty.js?xhr is in the entries buffer");
    });
</script>
</body>
</html>
